Don't use callback style API and switch to mpsc::Sender/Receiver. Also add lock around take().
The callback, while similar to C's API was inconvenient in Rust at best.
Implementing a BufferGuard similar to MutexGuard allows us to send the
buffer around and still tidy up afterwords.

Importantly, this should now be safe with the addition of the lock in
take(). As a bonus, performance is also vastly improved and testing
suggests over 392 JPEGs per second compared to around 60 with
the callback.